feat: Consolidate type formatting and validation into types module
Move example formatting and number rendering functions from metadata.rs
to types.rs, centralizing all type-related functionality in a single module.
This refactoring improves code organization and makes the TypeRef and
RustPrimitive types more self-contained.

Changes:
- Move format_example_value and related helpers to TypeRef::format_example()
- Move format_to_primitive to RustPrimitive::from_format()
- Move render_number, render_integer, render_unsigned_integer to RustPrimitive methods
- Move date/time parsing and formatting helpers to types.rs
- Move string escaping utilities to types.rs
- Update extract_validation_attrs to accept TypeRef for proper type context
- Update all call sites to use new API
