+++ crates/oas3-gen/src/generator/converter/operations.rs
@@ -486,7 +486,7 @@ impl<'a> OperationConverter<'a> {
     let schema = schema_ref.resolve(self.spec)?;
     let type_ref = self.schema_converter.schema_to_type_ref(&schema)?;
     let is_required = param.required.unwrap_or(false);
-    let validation = SchemaConverter::extract_validation_attrs(&param.name, is_required, &schema);
+    let validation = SchemaConverter::extract_validation_attrs(&param.name, is_required, &schema, &type_ref);
     let regex = SchemaConverter::extract_validation_pattern(&param.name, &schema).cloned();
     let default = SchemaConverter::extract_default_value(&schema);
